requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# LLM Integration
openai==1.3.0
ollama>=0.1.5

# Presentation Generation
python-pptx==0.6.23
google-api-python-client==2.100.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0

# Data Validation
pydantic>=2.9.0
pydantic-settings>=2.5.0

# Utilities
python-dotenv==1.0.0
loguru==0.7.2
aiofiles==23.2.1

# Image Processing (optional)
Pillow>=10.2.0
requests==2.31.0

# Image Search
duckduckgo_search>=4.0.0
